refactor(components): merge getDate \& useForecastRender, useCurrentIcon \& TheCurrentForecast
МЕТА ДАННЫЕ: refactor 11
