+++ src/components/UI/TheFavoriteCity.vue
@@ -32,7 +32,6 @@
 </template>
 
 <script>
-import useFirstUpperCase from '@/hooks/useFirstUpperCase';
 import useFetchWeather from '@/hooks/useFetchWeather';
 import useFavoriteLocation from '@/hooks/useFavoriteLocation';
 
@@ -45,13 +44,11 @@ export default {
     },
   },
   setup(props) {
-    const { firstUpperCase } = useFirstUpperCase();
     const { imgSrc, temperature, description, fetchCityWeather } =
-      useFetchWeather(firstUpperCase);
+      useFetchWeather();
     const { removeCity } = useFavoriteLocation();
 
     return {
-      firstUpperCase,
       imgSrc,
       temperature,
       description,
