+++ src/components/UI/TheSearchHistoryItem.vue
@@ -17,7 +17,6 @@
 <script setup>
 import { toRefs } from 'vue';
 import useFetchWeather from '@/hooks/useFetchWeather';
-import useFirstUpperCase from '@/hooks/useFirstUpperCase';
 
 const props = defineProps({
   historyItem: {
@@ -27,9 +26,7 @@ const props = defineProps({
 });
 
 const { historyItem } = toRefs(props);
-const { firstUpperCase } = useFirstUpperCase();
-const { imgSrc, tempMin, tempMax, fetchCityWeather } =
-  useFetchWeather(firstUpperCase);
+const { imgSrc, tempMin, tempMax, fetchCityWeather } = useFetchWeather();
 </script>
 
 <style>
